refactor(gfql): type + relocate connected-join pushdown internals
Type/architecture cleanup from code review of the connected-join predicate
pushdown. All changes are typing or relocation -- zero runtime-behavior change
(1336 cypher+filter_by_dict tests pass; connected-join answers unchanged).

- Operators: retype the pushdown gate's `op: str` to the existing `WhereOp`
  Literal (ast.py); narrow `_connected_join_literal_op` -> `Optional[WhereOp]`;
  delete both `cast(str, predicate.op)` widenings (predicate.op is already
  WhereOp). value/dtype/node_dtypes params typed via new aliases.
- Aliases: add `DType` (engine-polymorphic column dtype, honestly Any) and
  `NodeDtypes = Mapping[str, DType]` to compute/typing.py; use across the gate,
  gfql_unified, and the relocated helpers.
- Relocation: move `_LazyNodeDtypes`, `_node_dtypes_for_pushdown`,
  `_object_column_holds_non_strings`, `_read_node_dtypes` from the gfql_unified
  dispatch file to compute/filter_by_dict.py -- their true home (it owns the
  `_is_*_dtype_safe` validators these use and already imports the engine deps).
  No circular import. Test imports + the `_read_node_dtypes` monkeypatch
  retarget to filter_by_dict.
- Dynamic access: `getattr(g, "_nodes", None)` -> `g._nodes` (Plottable._nodes
  is a declared attribute) in the relocated helpers.
- Deprecated path: `_compile_string_query` now calls `compile_cypher_query`
  directly instead of threading a private `_node_dtypes` through the deprecated
  `compile_cypher`; that param is removed, restoring compile_cypher's compat
  surface (api.py now identical to master).
- Surface-guard baseline bumped 9235 -> 9237 for the two new import lines.
