+++ graphistry/compute/typing.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from typing import Any, Protocol, TYPE_CHECKING, Tuple, TypeVar, Union
+from typing import Any, Mapping, Protocol, TYPE_CHECKING, Tuple, TypeVar, Union
 
 # TODO stubs for Union[cudf.DataFrame, dask.DataFrame, ..] at checking time
 if TYPE_CHECKING:
@@ -12,6 +12,12 @@ else:
     SeriesT = Any
     IndexT = Any
     DomainT = Any
+
+# Engine-polymorphic column dtype: numpy dtype / pandas ExtensionDtype / polars DataType.
+# Honestly Any -- the concrete type is engine-dependent and only ever passed to dtype-inspection
+# helpers that accept Any and fail closed.
+DType = Any
+NodeDtypes = Mapping[str, DType]
 
 # Type variable for return type preservation in predicates
 T = TypeVar('T')
